tests/wpt/web-platform-tests/encoding-detection/el-ISO-8859-7-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>el ISO-8859-7 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "ISO-8859-7", 'Expected ISO-8859-7');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/el-ISO-8859-7-late.sub.html");
</script>
